Back port of the Yamagi Quake 2 save system
Major difference is that instead of recording pointers to functions or mmove definitions, it records the pointer name. This allows it to do a lookup of the actual address at load time, making it fairly immune to differences in memory layouts from builds. The only thing that will cause it to break is significant changes to the fields it actually stores.

This is likely buggy for the moment - there is no easy way to tell if things are correct without loading it and trying to save, but it appears (mostly) stable.